Add assessment utility aliases
