+++ tests/components/StudentTestForm.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
 import { fireEvent, render, screen, waitFor, within } from '@testing-library/react'
 import { StudentTestForm } from '@/components/StudentTestForm'
-import { createMockQuizQuestion } from '../helpers/mocks'
+import { createMockTestQuestion } from '../helpers/mocks'
 
 describe('StudentTestForm preview mode', () => {
   beforeEach(() => {
@@ -20,7 +20,7 @@ describe('StudentTestForm preview mode', () => {
       <StudentTestForm
         quizId="legacy-test-id"
         questions={[
-          createMockQuizQuestion({
+          createMockTestQuestion({
             id: 'q1',
             question_text: 'Which option is correct?',
             options: ['A', 'B'],
@@ -47,7 +47,7 @@ describe('StudentTestForm preview mode', () => {
       <StudentTestForm
         testId="test-preview-id"
         questions={[
-          createMockQuizQuestion({
+          createMockTestQuestion({
             id: 'q1',
             question_text: 'Which option is correct?',
             options: ['A', 'B'],
@@ -78,7 +78,7 @@ describe('StudentTestForm preview mode', () => {
       <StudentTestForm
         testId="test-open-response-label-id"
         questions={[
-          createMockQuizQuestion({
+          createMockTestQuestion({
             id: 'q1',
             question_text: 'Explain your reasoning.',
             options: [],
@@ -101,7 +101,7 @@ describe('StudentTestForm preview mode', () => {
       <StudentTestForm
         testId="test-flag-id"
         questions={[
-          createMockQuizQuestion({
+          createMockTestQuestion({
             id: 'q1',
             question_text: 'Which option is correct?',
             options: ['A', 'B'],
@@ -144,7 +144,7 @@ describe('StudentTestForm preview mode', () => {
       <StudentTestForm
         testId="test-warning-id"
         questions={[
-          createMockQuizQuestion({
+          createMockTestQuestion({
             id: 'q1',
             question_text: 'Which option is correct?',
             options: ['A', 'B'],
@@ -181,14 +181,14 @@ describe('StudentTestForm preview mode', () => {
       <StudentTestForm
         testId="test-footer-id"
         questions={[
-          createMockQuizQuestion({
+          createMockTestQuestion({
             id: 'q1',
             question_text: 'Question 1?',
             options: ['A', 'B'],
             question_type: 'multiple_choice',
             position: 0,
           }),
-          createMockQuizQuestion({
+          createMockTestQuestion({
             id: 'q2',
             question_text: 'Question 2?',
             options: ['C', 'D'],
@@ -226,7 +226,7 @@ describe('StudentTestForm preview mode', () => {
       <StudentTestForm
         testId="test-radio-position-id"
         questions={[
-          createMockQuizQuestion({
+          createMockTestQuestion({
             id: 'q1',
             question_text: 'Which option is correct?',
             options: ['A', 'B'],
@@ -256,7 +256,7 @@ describe('StudentTestForm preview mode', () => {
       <StudentTestForm
         testId="test-markdown-options-id"
         questions={[
-          createMockQuizQuestion({
+          createMockTestQuestion({
             id: 'q1',
             question_text: 'Which code compiles?',
             options: [
@@ -291,7 +291,7 @@ describe('StudentTestForm preview mode', () => {
       <StudentTestForm
         testId="test-closed-id"
         questions={[
-          createMockQuizQuestion({
+          createMockTestQuestion({
             id: 'q1',
             question_text: 'Which option is correct?',
             options: ['A', 'B'],
